% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allScaffolds.R
\docType{data}
\name{allScaffolds}
\alias{allScaffolds}
\title{All scaffolds}
\format{
a data.frame
}
\source{
/inst/script/allScaffolds.R
}
\usage{
data("allScaffolds", package = "spaceRATScaffolds")
}
\value{
a data.frame
}
\description{
This is a dataframe of all available scaffolds
}
\keyword{datasets}
