% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilaria.R
\docType{data}
\name{counts_ilaria}
\alias{counts_ilaria}
\title{Expression matrix as new samples}
\format{
a data frame
}
\source{
Kindly offered by Fred
}
\usage{
data("exprs_ilaria", package = "spaceRATScaffolds")
}
\value{
a matrix
}
\description{
counts_ilaria contains counts matrix of
Acute Myeloid Leukemia of erythroid subtype.
}
\keyword{datasets}
