% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmap.R
\docType{data}
\name{DMAP.v1}
\alias{DMAP.v1}
\title{DMAP scaffold version 1}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24759}
}
\value{
a spaceRAT scaffold
}
\description{
Scaffold of normal hematopoietic cells. 
The data was measured by microarray. \cr
\cr
Blood Cell Type Code Reference \cr
\itemize{
  \item BASO: Basophil
  \item BASO: Basophil 
  \item CMP:  Common Myeloid Progenitor
  \item EOS:  Eosinophil 
  \item ERY:  Erythrocyte 
  \item GMP:  Granulocyte-Macrophage Progenitor
  \item GRAN: Granulocyte
  \item HSC:  Hematopoietic Stem Cell 
  \item MEGA: Megakaryocyte 
  \item MEP:  Megakaryocyte-Erythroid Progenitor
  \item MONO: Monocyte
  }
}
\keyword{datasets}
