% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmap.R
\docType{data}
\name{DMAP.v1_scaffold}
\alias{DMAP.v1_scaffold}
\title{DMAP scaffold}
\format{
a scaffoldSpace
}
\source{
Created from exprs_dmap
}
\usage{
getScaffold("DMAP")
}
\value{
a scaffoldSpace
}
\description{
This scaffold was build from the
\code{\link{exprs_dmap}} object.
}
\keyword{datasets}
