% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmap.R
\docType{data}
\name{DMAP_exprs}
\alias{DMAP_exprs}
\title{Expression matrix for scaffold}
\format{
a matrix
}
\source{
Kindly offered by Frederik Otzen Bagger
}
\usage{
data("DMAP_exprs", package = "spaceRATScaffolds")
}
\value{
a matrix
}
\description{
exprs_map records expression matrix of normal
hematopoietic cells. The data is measured by microarray.
}
\keyword{datasets}
