% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmap.R
\docType{data}
\name{DMAP_pData}
\alias{DMAP_pData}
\title{Phenotype data for scaffold}
\format{
a data frame
}
\source{
Kindly offered by Frederik Otzen Bagger
}
\usage{
data("DMAP_pData", package = "spaceRATScaffolds")
}
\value{
a data.frame
}
\description{
pData_map records phenotype information
corresponding to DMAP_exprs
}
\keyword{datasets}
