% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_converters.R
\docType{data}
\name{gene_id_converter_hs}
\alias{gene_id_converter_hs}
\title{Gene ID converter for human}
\format{
a data frame
}
\source{
downloaded from biomaRt
}
\usage{
data("gene_id_converter_hs", package = "spaceRATScaffolds")
}
\value{
a data.frame
}
\description{
This table contains ensembl_gene_id of all human genes,
and corresponding ensembl_transcript_id, entrezgene_id ,hgnc_symbol,
and refseq_mrna
}
\keyword{datasets}
