% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_converters.R
\docType{data}
\name{gene_id_converter_mm}
\alias{gene_id_converter_mm}
\title{Gene ID converter for Mus musculus (mouse)}
\format{
a data frame
}
\source{
downloaded from biomaRt
}
\usage{
data("gene_id_converter_mm", package = "spaceRATScaffolds")
}
\value{
a data.frame
}
\description{
This table contains ensembl_gene_id of all mouse genes,
}
\keyword{datasets}
