% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allScaffolds.R
\name{getAllScaffolds}
\alias{getAllScaffolds}
\title{Get data frame of all scaffolds uploaded to Zenodo}
\usage{
getAllScaffolds(doi = "10.5281/zenodo.10142340")
}
\arguments{
\item{doi}{The DOI of the Zenodo entry}
}
\value{
A data.frame of scaffolds
}
\description{
Get data frame of all scaffolds uploaded to Zenodo
}
\examples{
allScaffolds <- getAllScaffolds()
}
