% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaffold.R
\name{getScaffold}
\alias{getScaffold}
\title{Helper function to get spaceRAT scaffolds}
\usage{
getScaffold(name, store = FALSE, path = "scaffolds", timeout = 120)
}
\arguments{
\item{name}{Name of scaffold to get}

\item{store}{Whether or not to store the scaffold for future use}

\item{path}{Path to folder for storing scaffolds}

\item{timeout}{Setting download timeout}
}
\value{
A spaceRAT scaffold
}
\description{
This function acquires a given prebuilt scaffold from the scaffolds
available on Zenodo. To prevent repeated downloads, `store` can be set to
`TRUE`. This will create a "scaffolds" folder, where the scaffolds are
stored.
}
\examples{
scaffold <- getScaffold("DMAP")
}
