% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaffold.R
\name{getZenodo}
\alias{getZenodo}
\title{Extract Zenodo information about a scaffold}
\usage{
getZenodo(scaffoldName)
}
\arguments{
\item{scaffoldName}{Name of scaffold}
}
\value{
A data.frame with Zenodo location of scaffold
}
\description{
Extract Zenodo information about a scaffold
}
\examples{
zenodo <- getZenodo("DMAP")
}
