% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\docType{data}
\name{GTEx.v1}
\alias{GTEx.v1}
\title{GTEx scaffold version 1}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://www.gtexportal.org/home/}
}
\value{
a spaceRAT scaffold
}
\description{
A spaceRAT scaffold of The Genotype-Tissue Expression project.
}
\keyword{datasets}
