% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilaria.R
\docType{data}
\name{ilaria_counts}
\alias{ilaria_counts}
\title{Counts matrix as new samples}
\format{
a data frame
}
\source{
Kindly offered by Frederik Otzen Bagger
}
\usage{
data("ilaria_counts", package = "spaceRATScaffolds")
}
\value{
a matrix
}
\description{
ilaria_counts contains counts matrix of
Acute Myeloid Leukemia of erythroid subtype.
}
\keyword{datasets}
