% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilaria.R
\docType{data}
\name{ilaria_pData}
\alias{ilaria_pData}
\title{Phenotype data for new samples}
\format{
a data frame
}
\source{
Kindly offered by Frederik Otzen Bagger
}
\usage{
data("ilaria_pData", package = "spaceRATScaffolds")
}
\value{
a data.frame
}
\description{
ilaria_pData contains phenotype
data corresponding to counts_ilaria
}
\keyword{datasets}
