% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listScaffolds.R
\name{listConverters}
\alias{listConverters}
\title{List available ID converters}
\usage{
listConverters()
}
\value{
A list of ID converters names
}
\description{
This function will list the gene ID converters available in
`spaceRATscaffolds`.
}
\examples{
listConverters()
}
