% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listScaffolds.R
\name{listDatasets}
\alias{listDatasets}
\title{List available datasets}
\usage{
listDatasets()
}
\value{
A list of datasets names
}
\description{
This function will list the test datasets available in `spaceRATScaffolds`.
}
\examples{
listDatasets()
}
