% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listScaffolds.R
\name{listScaffolds}
\alias{listScaffolds}
\title{List available scaffolds}
\usage{
listScaffolds()
}
\value{
A list of scaffold names
}
\description{
List available scaffolds
}
\examples{
listScaffolds()
}
