% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Helper function to load package data}
\usage{
loadData(name)
}
\value{
a spaceRATScaffolds data object
}
\description{
Helper function to load package data
}
\examples{
testData <- loadData("DMAP_exprs")
}
