% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\docType{data}
\name{TCGA.v2}
\alias{TCGA.v2}
\title{TCGA scaffold version 2}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://portal.gdc.cancer.gov/}
}
\value{
a spaceRAT scaffold
}
\description{
A spaceRAT scaffold of The Cancer Genome Atlas.
}
\keyword{datasets}
