% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmap.R
\docType{data}
\name{DMAP.v1}
\alias{DMAP.v1}
\title{DMAP scaffold version 1}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24759}
}
\value{
a spaceRAT scaffold
}
\description{
Scaffold of normal hematopoietic cells.
The data was measured by microarray.
**Blood Cell Type Code Reference**
BASO: Basophil
CMP:  Common Myeloid Progenitor
EOS:  Eosinophil
ERY:  Erythrocyte
GMP:  Granulocyte-Macrophage Progenitor
GRAN: Granulocyte
HSC:  Hematopoietic Stem Cell
MEGA: Megakaryocyte
MEP:  Megakaryocyte-Erythroid Progenitor
MONO: Monocyte
}
\keyword{datasets}
