% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaffold.R
\name{getScaffold}
\alias{getScaffold}
\title{Helper function to get spaceRAT scaffolds}
\usage{
getScaffold(name, store = TRUE, path = "scaffolds", timeout = 120)
}
\arguments{
\item{name}{Name of scaffold to get}

\item{store}{Whether or not to store the scaffold for future use}

\item{path}{Path to folder for storing scaffolds}

\item{timeout}{Setting download timeout}
}
\description{
Helper function to get spaceRAT scaffolds
}
