% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolds.R
\name{listConverters}
\alias{listConverters}
\title{List available ID converters}
\usage{
listConverters()
}
\value{
A list of ID converters names
}
\description{
List available ID converters
}
\examples{
listConverters()
}
