% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listScaffolds.R
\name{listScaffolds}
\alias{listScaffolds}
\title{List available scaffolds}
\usage{
listScaffolds()
}
\value{
A list of scaffold names
}
\description{
This function will check the prebuilt scaffolds available on Zendodo and
list them.
}
\examples{
listScaffolds()
}
