% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Helper function to load package data}
\usage{
loadData(name)
}
\arguments{
\item{name}{Name of object to load}
}
\value{
a spaceRATScaffolds data object
}
\description{
This is a helper function that extracts a dataset in `spaceRATScaffolds`,
allowing the user to store the object in the name of their choice.
}
\examples{
testData <- loadData("DMAP_exprs")
}
