% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\docType{data}
\name{TCGA.v2}
\alias{TCGA.v2}
\title{TCGA scaffold version 2}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://portal.gdc.cancer.gov/}
}
\value{
a spaceRAT scaffold
}
\description{
A spaceRAT scaffold of The Cancer Genome Atlas. \cr
\cr
TCGA Cancer Type Code Reference \cr
\itemize{
  \item ACC: Adrenocortical carcinoma
  \item BLCA: Bladder Urothelial Carcinoma
  \item BRCA: Breast invasive carcinoma
  \item CESC: Cervical squamous cell carcinoma and endocervical adenocarcinoma
  \item CHOL: Cholangiocarcinoma
  \item COAD: Colon adenocarcinoma
  \item DLBC: Lymphoid Neoplasm Diffuse Large B-cell Lymphoma
  \item ESCA: Esophageal carcinoma
  \item GBM: Glioblastoma multiforme
  \item HNSC: Head and Neck squamous cell carcinoma
  \item KICH: Kidney Chromophobe
  \item KIRC: Kidney renal clear cell carcinoma
  \item KIRP: Kidney renal papillary cell carcinoma
  \item LAML: Acute Myeloid Leukemia
  \item LGG: Brain Lower Grade Glioma
  \item LIHC: Liver hepatocellular carcinoma
  \item LUAD: Lung adenocarcinoma
  \item LUSC: Lung squamous cell carcinoma
  \item MESO: Mesothelioma
  \item OV: Ovarian serous cystadenocarcinoma
  \item PAAD: Pancreatic adenocarcinoma
  \item PCPG: Pheochromocytoma and Paraganglioma
  \item PRAD: Prostate adenocarcinoma
  \item READ: Rectum adenocarcinoma
  \item SARC: Sarcoma
  \item SKCM: Skin Cutaneous Melanoma
  \item STAD: Stomach adenocarcinoma
  \item TGCT: Testicular Germ Cell Tumors
  \item THCA: Thyroid carcinoma
  \item THYM: Thymoma
  \item UCEC: Uterine Corpus Endometrial Carcinoma
  \item UCS: Uterine Carcinosarcoma
  \item UVM: Uveal Melanoma
}
}
\keyword{datasets}
