% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\docType{data}
\name{TCGA.v2}
\alias{TCGA.v2}
\title{TCGA scaffold version 2}
\format{
a spaceRAT scaffold
}
\source{
Raw data available at:
\url{https://portal.gdc.cancer.gov/}
}
\value{
a spaceRAT scaffold
}
\description{
A spaceRAT scaffold of The Cancer Genome Atlas.
**TCGA Cancer Type Code Reference**
ACC:  Adrenocortical carcinoma
BLCA: Bladder Urothelial Carcinoma
BRCA: Breast invasive carcinoma
CESC: Cervical squamous cell carcinoma and endocervical adenocarcinoma
CHOL: Cholangiocarcinoma
COAD: Colon adenocarcinoma
DLBC: Lymphoid Neoplasm Diffuse Large B-cell Lymphoma
ESCA: Esophageal carcinoma
GBM:  Glioblastoma multiforme
HNSC: Head and Neck squamous cell carcinoma
KICH: Kidney Chromophobe
KIRC: Kidney renal clear cell carcinoma
KIRP: Kidney renal papillary cell carcinoma
LAML: Acute Myeloid Leukemia
LGG:  Brain Lower Grade Glioma
LIHC: Liver hepatocellular carcinoma
LUAD: Lung adenocarcinoma
LUSC: Lung squamous cell carcinoma
MESO: Mesothelioma
OV:   Ovarian serous cystadenocarcinoma
PAAD: Pancreatic adenocarcinoma
PCPG: Pheochromocytoma and Paraganglioma
PRAD: Prostate adenocarcinoma
READ: Rectum adenocarcinoma
SARC: Sarcoma
SKCM: Skin Cutaneous Melanoma
STAD: Stomach adenocarcinoma
TGCT: Testicular Germ Cell Tumors
THCA: Thyroid carcinoma
THYM: Thymoma
UCEC: Uterine Corpus Endometrial Carcinoma
UCS:  Uterine Carcinosarcoma
UVM:  Uveal Melanoma
}
\keyword{datasets}
